+++ src/app/activities/chief-of-clinicians/page.tsx
@@ -1,1186 +1,8 @@
-import InstructorActivitiesContent from "../clinical-instructor/InstructorsActivities"
-// import ActivitiesPage from "./ActivitiesPage"
+import UnifiedActivitiesRecords from "@/app/records/Records"
 
 export default function AdminActivitiesPage() {
   return (
-    <InstructorActivitiesContent />
-    // <div className="space-y-6">
-    //   <h1 className="text-2xl font-bold text-[#333]">Activities & Grades Management</h1>
-    //   <p className="text-[#666]">View, manage, and grade activities assigned to clinicians under your supervision.</p>
-      
-    // </div>
+    <UnifiedActivitiesRecords />
   )
 }
-// "use client"
 
-// import { useState } from "react"
-// import {
-//   ArrowUpDown,
-//   Check,
-//   ChevronDown,
-//   ChevronUp,
-//   Download,
-//   Edit,
-//   Plus,
-//   X,
-//   AlertCircle,
-//   CheckCircle,
-// } from "lucide-react"
-// import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-// import { Button } from "@/components/ui/button"
-// import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
-// import { Input } from "@/components/ui/input"
-// import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from "@/components/ui/dialog"
-// import { Label } from "@/components/ui/label"
-// import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-// import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-// import { Alert, AlertDescription } from "@/components/ui/alert"
-
-// // Define types directly in the component file
-// interface HistoryEntry {
-//   timestamp: string
-//   user: string
-//   action: "Created" | "Updated" | "Archived"
-//   description: string
-//   details: {
-//     date: string
-//     procedure: string
-//     patient: string
-//     grade: string
-//     remarks: string
-//   }
-// }
-
-// interface Activity {
-//   id: string
-//   firstName: string
-//   lastName: string
-//   chair: string
-//   patient: string
-//   instructor: string
-//   procedure: string
-//   procedures?: string[]
-//   status: "Not started" | "Started" | "Completed" | "Incomplete"
-//   date: string
-//   grade?: string
-//   remarks?: string
-//   assessmentStatus?: "in-progress" | "graded"
-//   archived?: boolean // Add archived field
-//   history: HistoryEntry[]
-// }
-
-// interface NewActivityForm {
-//   firstName: string
-//   lastName: string
-//   chair: string
-//   patient: string
-//   instructor: string
-//   procedures: string[]
-//   status: Activity["status"]
-// }
-
-// export default function ActivitiesPage() {
-//   const [isModalOpen, setIsModalOpen] = useState(false)
-//   const [isEditModalOpen, setIsEditModalOpen] = useState(false)
-//   const [currentActivity, setCurrentActivity] = useState<Activity | null>(null)
-//   const [isCompleteModalOpen, setIsCompleteModalOpen] = useState(false)
-//   const [isArchiveModalOpen, setIsArchiveModalOpen] = useState(false)
-//   const [activityToAction, setActivityToAction] = useState<Activity | null>(null)
-//   const [sortField, setSortField] = useState<"id" | "lastName">("id")
-//   const [sortDirection, setSortDirection] = useState<"asc" | "desc">("asc")
-//   const [isHistoryModalOpen, setIsHistoryModalOpen] = useState(false)
-//   const [selectedActivity, setSelectedActivity] = useState<Activity | null>(null)
-//   const [editFormData, setEditFormData] = useState<Partial<Activity>>({})
-
-//   // Form state for new activity
-//   const [newActivityForm, setNewActivityForm] = useState<NewActivityForm>({
-//     firstName: "",
-//     lastName: "",
-//     chair: "",
-//     patient: "",
-//     instructor: "",
-//     procedures: [],
-//     status: "Not started",
-//   })
-
-//   // Error and success states
-//   const [formErrors, setFormErrors] = useState<Record<string, string>>({})
-//   const [showSuccessAlert, setShowSuccessAlert] = useState(false)
-//   const [showErrorAlert, setShowErrorAlert] = useState(false)
-//   const [alertMessage, setAlertMessage] = useState("")
-
-//   // Sample data for all activities (including past ones)
-//   const [activities, setActivities] = useState<Activity[]>([
-//     {
-//       id: "1",
-//       firstName: "Maria",
-//       lastName: "Santos",
-//       chair: "Chair 05",
-//       patient: "Juan Dela Cruz",
-//       instructor: "Dr. Reyes",
-//       procedure: "Root Canal Treatment",
-//       procedures: ["Root Canal Treatment", "Dental Filling"],
-//       status: "Started",
-//       date: "2025-05-08",
-//       archived: false,
-//       history: [
-//         {
-//           timestamp: "2025-05-08T14:20:00",
-//           user: "admin@example.com",
-//           action: "Created",
-//           description: "Initial activity record created",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Root Canal Treatment",
-//             patient: "Juan Dela Cruz",
-//             grade: "",
-//             remarks: "",
-//           },
-//         },
-//       ],
-//     },
-//     {
-//       id: "2",
-//       firstName: "John",
-//       lastName: "Dela Cruz",
-//       chair: "Chair 12",
-//       patient: "Ana Reyes",
-//       instructor: "Dr. Mendoza",
-//       procedure: "Dental Filling",
-//       status: "Not started",
-//       date: "2025-05-08",
-//       archived: false,
-//       history: [
-//         {
-//           timestamp: "2025-05-08T10:15:00",
-//           user: "admin@example.com",
-//           action: "Created",
-//           description: "Initial activity record created",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Dental Filling",
-//             patient: "Ana Reyes",
-//             grade: "",
-//             remarks: "",
-//           },
-//         },
-//       ],
-//     },
-//     {
-//       id: "3",
-//       firstName: "Anna",
-//       lastName: "Lim",
-//       chair: "Chair 03",
-//       patient: "Miguel Santos",
-//       instructor: "Dr. Santos",
-//       procedure: "Dental Crown",
-//       procedures: ["Dental Crown", "Teeth Cleaning"],
-//       status: "Started",
-//       date: "2025-05-08",
-//       archived: false,
-//       history: [
-//         {
-//           timestamp: "2025-05-08T09:30:00",
-//           user: "admin@example.com",
-//           action: "Created",
-//           description: "Initial activity record created",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Dental Crown",
-//             patient: "Miguel Santos",
-//             grade: "",
-//             remarks: "",
-//           },
-//         },
-//       ],
-//     },
-//     {
-//       id: "4",
-//       firstName: "Mark",
-//       lastName: "Aquino",
-//       chair: "Chair 08",
-//       patient: "Sofia Reyes",
-//       instructor: "Dr. Reyes",
-//       procedure: "Teeth Cleaning",
-//       status: "Completed",
-//       date: "2025-05-08",
-//       grade: "88",
-//       remarks: "Thorough cleaning, good patient management",
-//       archived: false,
-//       history: [
-//         {
-//           timestamp: "2025-05-08T16:45:00",
-//           user: "dr.reyes@example.com",
-//           action: "Updated",
-//           description: "Status changed to Completed and grade added",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Teeth Cleaning",
-//             patient: "Sofia Reyes",
-//             grade: "88",
-//             remarks: "Thorough cleaning, good patient management",
-//           },
-//         },
-//         {
-//           timestamp: "2025-05-08T11:20:00",
-//           user: "admin@example.com",
-//           action: "Created",
-//           description: "Initial activity record created",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Teeth Cleaning",
-//             patient: "Sofia Reyes",
-//             grade: "",
-//             remarks: "",
-//           },
-//         },
-//       ],
-//     },
-//     {
-//       id: "5",
-//       firstName: "Sarah",
-//       lastName: "Garcia",
-//       chair: "Chair 10",
-//       patient: "Luis Tan",
-//       instructor: "Dr. Tan",
-//       procedure: "Dental Extraction",
-//       status: "Incomplete",
-//       date: "2025-05-08",
-//       archived: false,
-//       history: [
-//         {
-//           timestamp: "2025-05-08T17:30:00",
-//           user: "dr.tan@example.com",
-//           action: "Updated",
-//           description: "Status changed to Incomplete",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Dental Extraction",
-//             patient: "Luis Tan",
-//             grade: "",
-//             remarks: "Patient had to reschedule due to emergency",
-//           },
-//         },
-//         {
-//           timestamp: "2025-05-08T13:10:00",
-//           user: "admin@example.com",
-//           action: "Created",
-//           description: "Initial activity record created",
-//           details: {
-//             date: "2025-05-08",
-//             procedure: "Dental Extraction",
-//             patient: "Luis Tan",
-//             grade: "",
-//             remarks: "",
-//           },
-//         },
-//       ],
-//     },
-//   ])
-
-//   // Function to get status color
-//   const getStatusColor = (status: Activity["status"]): string => {
-//     switch (status) {
-//       case "Started":
-//         return "bg-[#5C8E77]/10 text-[#5C8E77]"
-//       case "Completed":
-//         return "bg-blue-50 text-blue-700"
-//       case "Not started":
-//         return "bg-yellow-50 text-yellow-700"
-//       case "Incomplete":
-//         return "bg-red-50 text-red-700"
-//       default:
-//         return "bg-gray-50 text-gray-700"
-//     }
-//   }
-
-//   // Validation function
-//   const validateForm = (form: NewActivityForm): Record<string, string> => {
-//     const errors: Record<string, string> = {}
-//     if (!form.firstName.trim()) errors.firstName = "First name is required"
-//     if (!form.lastName.trim()) errors.lastName = "Last name is required"
-//     if (!form.chair) errors.chair = "Chair selection is required"
-//     if (!form.patient.trim()) errors.patient = "Patient name is required"
-//     if (!form.instructor) errors.instructor = "Instructor selection is required"
-//     if (form.procedures.length === 0) errors.procedures = "At least one procedure must be selected"
-//     if (form.procedures.length > 2) errors.procedures = "Maximum 2 procedures allowed"
-//     return errors
-//   }
-
-//   // Handle procedure checkbox change
-//   const handleProcedureChange = (procedure: string, checked: boolean) => {
-//     setNewActivityForm((prev: NewActivityForm) => {
-//       const newProcedures = checked
-//         ? [...prev.procedures, procedure]
-//         : prev.procedures.filter((p: string) => p !== procedure)
-//       return { ...prev, procedures: newProcedures }
-//     })
-//   }
-
-//   // Handle form submission
-//   const handleCreateActivity = () => {
-//     const errors = validateForm(newActivityForm)
-//     setFormErrors(errors)
-
-//     if (Object.keys(errors).length > 0) {
-//       setShowErrorAlert(true)
-//       setAlertMessage("Please fix the errors below before submitting.")
-//       setTimeout(() => setShowErrorAlert(false), 5000)
-//       return
-//     }
-
-//     // Generate new activity
-//     const newActivity: Activity = {
-//       id: (activities.length + 1).toString(),
-//       firstName: newActivityForm.firstName,
-//       lastName: newActivityForm.lastName,
-//       chair: newActivityForm.chair,
-//       patient: newActivityForm.patient,
-//       instructor: newActivityForm.instructor,
-//       procedure: newActivityForm.procedures[0], // Primary procedure
-//       procedures: newActivityForm.procedures,
-//       status: newActivityForm.status,
-//       date: new Date().toISOString().slice(0, 10),
-//       archived: false,
-//       history: [
-//         {
-//           timestamp: new Date().toISOString(),
-//           user: "admin@example.com",
-//           action: "Created",
-//           description: "Initial activity record created",
-//           details: {
-//             date: new Date().toISOString().slice(0, 10),
-//             procedure: newActivityForm.procedures[0],
-//             patient: newActivityForm.patient,
-//             grade: "",
-//             remarks: "",
-//           },
-//         },
-//       ],
-//     }
-
-//     setActivities((prev: Activity[]) => [...prev, newActivity])
-
-//     // Reset form
-//     setNewActivityForm({
-//       firstName: "",
-//       lastName: "",
-//       chair: "",
-//       patient: "",
-//       instructor: "",
-//       procedures: [],
-//       status: "Not started",
-//     })
-//     setFormErrors({})
-//     setIsModalOpen(false)
-
-//     setShowSuccessAlert(true)
-//     setAlertMessage(`Activity created successfully for ${newActivity.firstName} ${newActivity.lastName}`)
-//     // Hide success alert after 5 seconds
-//     setTimeout(() => setShowSuccessAlert(false), 5000)
-//   }
-
-//   // Function to handle edit button click
-//   const handleEdit = (activity: Activity) => {
-//     setCurrentActivity(activity)
-//     setEditFormData(activity)
-//     setIsEditModalOpen(true)
-//   }
-
-//   const handleCompleteClick = (activity: Activity) => {
-//     setActivityToAction(activity)
-//     setIsCompleteModalOpen(true)
-//   }
-
-//   const handleComplete = () => {
-//     if (activityToAction) {
-//       const newHistory = {
-//         timestamp: new Date().toISOString(),
-//         user: "admin@example.com",
-//         action: "Updated" as const,
-//         description: "Status changed to Completed",
-//         details: {
-//           date: activityToAction.date,
-//           procedure: activityToAction.procedure,
-//           patient: activityToAction.patient,
-//           grade: activityToAction.grade || "",
-//           remarks: activityToAction.remarks || "",
-//         },
-//       }
-
-//       setActivities((activities: Activity[]) =>
-//         activities.map((activity) =>
-//           activity.id === activityToAction.id
-//             ? {
-//                 ...activity,
-//                 status: "Completed" as const,
-//                 history: [newHistory, ...activity.history],
-//               }
-//             : activity,
-//         ),
-//       )
-//       setIsCompleteModalOpen(false)
-//     }
-//   }
-
-//   // Updated archive functionality
-//   const handleArchiveClick = (activity: Activity) => {
-//     setActivityToAction(activity)
-//     setIsArchiveModalOpen(true)
-//   }
-
-//   const handleArchive = () => {
-//     if (activityToAction) {
-//       const newHistory = {
-//         timestamp: new Date().toISOString(),
-//         user: "admin@example.com",
-//         action: "Archived" as const,
-//         description: "Activity archived",
-//         details: {
-//           date: activityToAction.date,
-//           procedure: activityToAction.procedure,
-//           patient: activityToAction.patient,
-//           grade: activityToAction.grade || "",
-//           remarks: activityToAction.remarks || "",
-//         },
-//       }
-
-//       setActivities((activities: Activity[]) =>
-//         activities.map((activity) =>
-//           activity.id === activityToAction.id
-//             ? {
-//                 ...activity,
-//                 archived: true,
-//                 history: [newHistory, ...activity.history],
-//               }
-//             : activity,
-//         ),
-//       )
-//       setIsArchiveModalOpen(false)
-//       setShowSuccessAlert(true)
-//       setAlertMessage(`Activity archived successfully for ${activityToAction.firstName} ${activityToAction.lastName}`)
-//       setTimeout(() => setShowSuccessAlert(false), 5000)
-//     }
-//   }
-
-//   // Function to update activity
-//   const handleUpdateActivity = () => {
-//     if (!currentActivity || !editFormData) return
-
-//     const updatedActivity = { ...currentActivity, ...editFormData }
-//     const newHistory = {
-//       timestamp: new Date().toISOString(),
-//       user: "admin@example.com",
-//       action: "Updated" as const,
-//       description: "Activity details updated",
-//       details: {
-//         date: updatedActivity.date,
-//         procedure: updatedActivity.procedure,
-//         patient: updatedActivity.patient,
-//         grade: updatedActivity.grade || "",
-//         remarks: updatedActivity.remarks || "",
-//       },
-//     }
-
-//     setActivities((activities: Activity[]) =>
-//       activities.map((activity) =>
-//         activity.id === updatedActivity.id
-//           ? {
-//               ...updatedActivity,
-//               history: [newHistory, ...activity.history],
-//             }
-//           : activity,
-//       ),
-//     )
-
-//     setIsEditModalOpen(false)
-//     setShowSuccessAlert(true)
-//     setAlertMessage(`Activity updated successfully for ${updatedActivity.firstName} ${updatedActivity.lastName}`)
-//     setTimeout(() => setShowSuccessAlert(false), 5000)
-//   }
-
-//   // Function to handle sorting
-//   const handleSort = (field: "id" | "lastName") => {
-//     if (sortField === field) {
-//       setSortDirection(sortDirection === "asc" ? "desc" : "asc")
-//     } else {
-//       setSortField(field)
-//       setSortDirection("asc")
-//     }
-//   }
-
-//   // Function to export activities to CSV
-//   const exportActivitiesToCSV = () => {
-//     // Only export non-archived activities
-//     const activeActivities = activities.filter((activity) => !activity.archived)
-
-//     const headers = [
-//       "ID",
-//       "Date",
-//       "First Name",
-//       "Last Name",
-//       "Patient",
-//       "Chair",
-//       "Instructor",
-//       "Procedure",
-//       "Status",
-//       "Grade",
-//       "Remarks",
-//     ].join(",")
-
-//     const rows = activeActivities.map(
-//       (activity) =>
-//         `"${activity.id}","${activity.date}","${activity.firstName}","${activity.lastName}","${activity.patient}","${
-//           activity.chair
-//         }","${activity.instructor}","${activity.procedure}","${activity.status}","${activity.grade || ""}","${
-//           activity.remarks || ""
-//         }"`,
-//     )
-
-//     const csv = [headers, ...rows].join("\n")
-//     const blob = new Blob([csv], { type: "text/csv;charset=utf-8;" })
-//     const url = URL.createObjectURL(blob)
-//     const link = document.createElement("a")
-//     link.setAttribute("href", url)
-//     link.setAttribute("download", `Activities_${new Date().toISOString().slice(0, 10)}.csv`)
-//     link.style.visibility = "hidden"
-//     document.body.appendChild(link)
-//     link.click()
-//     document.body.removeChild(link)
-//   }
-
-//   // Filter out archived activities and sort
-//   const sortedActivities = activities
-//     .filter((activity) => !activity.archived)
-//     .sort((a, b) => {
-//       if (sortField === "id") {
-//         return sortDirection === "asc"
-//           ? Number.parseInt(a.id) - Number.parseInt(b.id)
-//           : Number.parseInt(b.id) - Number.parseInt(a.id)
-//       } else if (sortField === "lastName") {
-//         return sortDirection === "asc" ? a.lastName.localeCompare(b.lastName) : b.lastName.localeCompare(a.lastName)
-//       }
-//       return 0
-//     })
-
-//   return (
-//     <div className="min-h-screen bg-[#f8f9fa] p-6">
-//       {/* Success Alert */}
-//       {showSuccessAlert && (
-//         <Alert className="mb-4 border-green-200 bg-green-50 relative">
-//           <CheckCircle className="h-4 w-4 text-green-600" />
-//           <AlertDescription className="text-green-800">{alertMessage}</AlertDescription>
-//           <Button
-//             variant="ghost"
-//             size="sm"
-//             className="absolute right-2 top-2 h-6 w-6 p-0"
-//             onClick={() => setShowSuccessAlert(false)}
-//           >
-//             <X className="h-4 w-4" />
-//           </Button>
-//         </Alert>
-//       )}
-
-//       {/* Error Alert */}
-//       {showErrorAlert && (
-//         <Alert className="mb-4 border-red-200 bg-red-50 relative">
-//           <AlertCircle className="h-4 w-4 text-red-600" />
-//           <AlertDescription className="text-red-800">{alertMessage}</AlertDescription>
-//           <Button
-//             variant="ghost"
-//             size="sm"
-//             className="absolute right-2 top-2 h-6 w-6 p-0"
-//             onClick={() => setShowErrorAlert(false)}
-//           >
-//             <X className="h-4 w-4" />
-//           </Button>
-//         </Alert>
-//       )}
-
-//       {/* Activities Table */}
-//       <Card className="bg-white border border-gray-200 shadow-sm mb-6">
-//         <CardHeader className="flex flex-row items-center justify-between pb-4 border-b border-gray-200">
-//           <div>
-//             <CardTitle className="text-xl font-semibold text-[#333]">All Activities</CardTitle>
-//           </div>
-//           <div className="flex items-center gap-2">
-//             <Button
-//               variant="outline"
-//               size="sm"
-//               className="flex items-center gap-1 border-[#5C8E77] text-[#5C8E77] bg-transparent"
-//               onClick={exportActivitiesToCSV}
-//             >
-//               <Download className="h-4 w-4" />
-//               Export CSV
-//             </Button>
-//             <Button
-//               className="bg-[#5C8E77] hover:bg-[#406E58] text-white border-none flex items-center gap-2"
-//               onClick={() => setIsModalOpen(true)}
-//             >
-//               <Plus className="h-4 w-4" /> New Activity
-//             </Button>
-//           </div>
-//         </CardHeader>
-//         <CardContent className="p-0">
-//           <Table>
-//             <TableHeader className="bg-white border-b border-gray-200">
-//               <TableRow className="hover:bg-white border-b-0">
-//                 <TableHead className="font-medium text-[#333] cursor-pointer" onClick={() => handleSort("id")}>
-//                   <div className="flex items-center">
-//                     Act ID
-//                     <ArrowUpDown className="ml-1 h-4 w-4" />
-//                     {sortField === "id" &&
-//                       (sortDirection === "asc" ? (
-//                         <ChevronUp className="ml-1 h-4 w-4" />
-//                       ) : (
-//                         <ChevronDown className="ml-1 h-4 w-4" />
-//                       ))}
-//                   </div>
-//                 </TableHead>
-//                 <TableHead className="font-medium text-[#333]">First Name</TableHead>
-//                 <TableHead className="font-medium text-[#333] cursor-pointer" onClick={() => handleSort("lastName")}>
-//                   <div className="flex items-center">
-//                     Last Name
-//                     <ArrowUpDown className="ml-1 h-4 w-4" />
-//                     {sortField === "lastName" &&
-//                       (sortDirection === "asc" ? (
-//                         <ChevronUp className="ml-1 h-4 w-4" />
-//                       ) : (
-//                         <ChevronDown className="ml-1 h-4 w-4" />
-//                       ))}
-//                   </div>
-//                 </TableHead>
-//                 <TableHead className="font-medium text-[#333]">Patient</TableHead>
-//                 <TableHead className="font-medium text-[#333]">Chair</TableHead>
-//                 <TableHead className="font-medium text-[#333]">Instructor</TableHead>
-//                 <TableHead className="font-medium text-[#333]">Procedure</TableHead>
-//                 <TableHead className="font-medium text-[#333]">Status</TableHead>
-//                 <TableHead className="font-medium text-[#333]">Action</TableHead>
-//               </TableRow>
-//             </TableHeader>
-//             <TableBody>
-//               {sortedActivities.length > 0 ? (
-//                 sortedActivities.map((activity) => (
-//                   <TableRow key={activity.id} className="hover:bg-gray-50 border-b border-gray-200">
-//                     <TableCell className="font-medium text-[#333]">{activity.id}</TableCell>
-//                     <TableCell className="text-[#333]">{activity.firstName}</TableCell>
-//                     <TableCell className="text-[#333]">{activity.lastName}</TableCell>
-//                     <TableCell className="text-[#333]">{activity.patient}</TableCell>
-//                     <TableCell className="text-[#333]">{activity.chair}</TableCell>
-//                     <TableCell className="text-[#333]">{activity.instructor}</TableCell>
-//                     <TableCell className="text-[#333]">
-//                       {activity.procedures ? (
-//                         <div className="space-y-1">
-//                           {activity.procedures.map((proc: string, index: number) => (
-//                             <div key={index} className="text-sm">
-//                               {proc}
-//                             </div>
-//                           ))}
-//                         </div>
-//                       ) : (
-//                         activity.procedure
-//                       )}
-//                     </TableCell>
-//                     <TableCell>
-//                       <div
-//                         className={`px-3 py-1 rounded-full text-sm inline-flex items-center justify-center font-medium ${getStatusColor(
-//                           activity.status,
-//                         )}`}
-//                       >
-//                         {activity.status}
-//                       </div>
-//                     </TableCell>
-//                     <TableCell>
-//                       <div className="flex items-center gap-2">
-//                         <Button
-//                           size="icon"
-//                           variant="ghost"
-//                           className="h-8 w-8 text-[#5C8E77] hover:bg-[#e6f7eb]"
-//                           onClick={() => handleEdit(activity)}
-//                         >
-//                           <Edit className="h-4 w-4" />
-//                         </Button>
-//                         <Button
-//                           size="icon"
-//                           variant="ghost"
-//                           className="h-8 w-8 text-blue-600 hover:bg-blue-50"
-//                           onClick={() => handleCompleteClick(activity)}
-//                         >
-//                           <Check className="h-4 w-4" />
-//                         </Button>
-//                         <Button
-//                           size="icon"
-//                           variant="ghost"
-//                           className="h-8 w-8 text-orange-600 hover:bg-orange-50"
-//                           onClick={() => handleArchiveClick(activity)}
-//                         >
-//                           <X className="h-4 w-4" />
-//                         </Button>
-//                       </div>
-//                     </TableCell>
-//                   </TableRow>
-//                 ))
-//               ) : (
-//                 <TableRow>
-//                   <TableCell colSpan={9} className="text-center py-12 text-gray-500">
-//                     No activities found.
-//                   </TableCell>
-//                 </TableRow>
-//               )}
-//             </TableBody>
-//           </Table>
-//         </CardContent>
-//       </Card>
-
-//       {/* New Activity Modal */}
-//       <Dialog open={isModalOpen} onOpenChange={setIsModalOpen}>
-//         <DialogContent className="sm:max-w-[500px] p-0 overflow-hidden rounded-lg">
-//           <DialogHeader className="bg-[#f8f9fa] px-6 py-4 border-b border-gray-200">
-//             <DialogTitle className="text-xl font-semibold text-[#5C8E77]">New Activity</DialogTitle>
-//           </DialogHeader>
-//           <div className="px-6 py-4">
-//             <div className="grid gap-4">
-//               <div className="grid grid-cols-2 gap-4">
-//                 <div className="space-y-2">
-//                   <Label htmlFor="firstName" className="text-[#333]">
-//                     First Name
-//                   </Label>
-//                   <Input
-//                     id="firstName"
-//                     placeholder="Enter first name"
-//                     className={`border-gray-300 ${formErrors.firstName ? "border-red-500" : ""}`}
-//                     value={newActivityForm.firstName}
-//                     onChange={(e) =>
-//                       setNewActivityForm((prev: NewActivityForm) => ({ ...prev, firstName: e.target.value }))
-//                     }
-//                   />
-//                   {formErrors.firstName && <p className="text-sm text-red-600">{formErrors.firstName}</p>}
-//                 </div>
-//                 <div className="space-y-2">
-//                   <Label htmlFor="lastName" className="text-[#333]">
-//                     Last Name
-//                   </Label>
-//                   <Input
-//                     id="lastName"
-//                     placeholder="Enter last name"
-//                     className={`border-gray-300 ${formErrors.lastName ? "border-red-500" : ""}`}
-//                     value={newActivityForm.lastName}
-//                     onChange={(e) =>
-//                       setNewActivityForm((prev: NewActivityForm) => ({ ...prev, lastName: e.target.value }))
-//                     }
-//                   />
-//                   {formErrors.lastName && <p className="text-sm text-red-600">{formErrors.lastName}</p>}
-//                 </div>
-//               </div>
-//               <div className="grid grid-cols-2 gap-4">
-//                 <div className="space-y-2">
-//                   <Label htmlFor="chair" className="text-[#333]">
-//                     Chair
-//                   </Label>
-//                   <Select
-//                     value={newActivityForm.chair}
-//                     onValueChange={(value) =>
-//                       setNewActivityForm((prev: NewActivityForm) => ({ ...prev, chair: value }))
-//                     }
-//                   >
-//                     <SelectTrigger id="chair" className={`border-gray-300 ${formErrors.chair ? "border-red-500" : ""}`}>
-//                       <SelectValue placeholder="Select chair" />
-//                     </SelectTrigger>
-//                     <SelectContent>
-//                       <SelectItem value="Chair 03">Chair 03</SelectItem>
-//                       <SelectItem value="Chair 05">Chair 05</SelectItem>
-//                       <SelectItem value="Chair 08">Chair 08</SelectItem>
-//                       <SelectItem value="Chair 10">Chair 10</SelectItem>
-//                       <SelectItem value="Chair 12">Chair 12</SelectItem>
-//                     </SelectContent>
-//                   </Select>
-//                   {formErrors.chair && <p className="text-sm text-red-600">{formErrors.chair}</p>}
-//                 </div>
-//                 <div className="space-y-2">
-//                   <Label htmlFor="instructor" className="text-[#333]">
-//                     Instructor
-//                   </Label>
-//                   <Select
-//                     value={newActivityForm.instructor}
-//                     onValueChange={(value) =>
-//                       setNewActivityForm((prev: NewActivityForm) => ({ ...prev, instructor: value }))
-//                     }
-//                   >
-//                     <SelectTrigger
-//                       id="instructor"
-//                       className={`border-gray-300 ${formErrors.instructor ? "border-red-500" : ""}`}
-//                     >
-//                       <SelectValue placeholder="Select instructor" />
-//                     </SelectTrigger>
-//                     <SelectContent>
-//                       <SelectItem value="Dr. Reyes">Dr. Reyes</SelectItem>
-//                       <SelectItem value="Dr. Mendoza">Dr. Mendoza</SelectItem>
-//                       <SelectItem value="Dr. Santos">Dr. Santos</SelectItem>
-//                       <SelectItem value="Dr. Tan">Dr. Tan</SelectItem>
-//                     </SelectContent>
-//                   </Select>
-//                   {formErrors.instructor && <p className="text-sm text-red-600">{formErrors.instructor}</p>}
-//                 </div>
-//               </div>
-//               <div className="space-y-2">
-//                 <Label htmlFor="patient" className="text-[#333]">
-//                   Patient
-//                 </Label>
-//                 <Input
-//                   id="patient"
-//                   placeholder="Enter patient name"
-//                   className={`border-gray-300 ${formErrors.patient ? "border-red-500" : ""}`}
-//                   value={newActivityForm.patient}
-//                   onChange={(e) =>
-//                     setNewActivityForm((prev: NewActivityForm) => ({ ...prev, patient: e.target.value }))
-//                   }
-//                 />
-//                 {formErrors.patient && <p className="text-sm text-red-600">{formErrors.patient}</p>}
-//               </div>
-//               <div className="space-y-2">
-//                 <Label htmlFor="procedure" className="text-[#333]">
-//                   Procedures (Select up to 2)
-//                 </Label>
-//                 <div className="grid grid-cols-2 gap-2">
-//                   {[
-//                     "Root Canal Treatment",
-//                     "Dental Filling",
-//                     "Dental Crown",
-//                     "Teeth Cleaning",
-//                     "Dental Extraction",
-//                   ].map((procedure) => (
-//                     <div key={procedure} className="flex items-center space-x-2">
-//                       <input
-//                         type="checkbox"
-//                         id={procedure.toLowerCase().replace(/\s+/g, "-")}
-//                         className="h-4 w-4 rounded border-gray-300 text-[#5C8E77] focus:ring-[#5C8E77]"
-//                         checked={newActivityForm.procedures.includes(procedure)}
-//                         onChange={(e) => handleProcedureChange(procedure, e.target.checked)}
-//                         disabled={
-//                           !newActivityForm.procedures.includes(procedure) && newActivityForm.procedures.length >= 2
-//                         }
-//                       />
-//                       <label htmlFor={procedure.toLowerCase().replace(/\s+/g, "-")} className="text-sm text-gray-700">
-//                         {procedure}
-//                       </label>
-//                     </div>
-//                   ))}
-//                 </div>
-//                 {formErrors.procedures && <p className="text-sm text-red-600">{formErrors.procedures}</p>}
-//                 <p className="text-xs text-gray-500">You can select up to 2 procedures per activity.</p>
-//               </div>
-//               <div className="space-y-2">
-//                 <Label htmlFor="status" className="text-[#333]">
-//                   Status
-//                 </Label>
-//                 <Select
-//                   value={newActivityForm.status}
-//                   onValueChange={(value: Activity["status"]) =>
-//                     setNewActivityForm((prev: NewActivityForm) => ({ ...prev, status: value }))
-//                   }
-//                 >
-//                   <SelectTrigger id="status" className="border-gray-300">
-//                     <SelectValue placeholder="Select status" />
-//                   </SelectTrigger>
-//                   <SelectContent>
-//                     <SelectItem value="Not started">Not started</SelectItem>
-//                     <SelectItem value="Started">Started</SelectItem>
-//                     <SelectItem value="Completed">Completed</SelectItem>
-//                     <SelectItem value="Incomplete">Incomplete</SelectItem>
-//                   </SelectContent>
-//                 </Select>
-//               </div>
-//             </div>
-//           </div>
-//           <DialogFooter className="bg-[#f8f9fa] px-6 py-4 border-t border-gray-200">
-//             <Button variant="outline" onClick={() => setIsModalOpen(false)} className="border-gray-300">
-//               Cancel
-//             </Button>
-//             <Button className="bg-[#5C8E77] hover:bg-[#406E58] text-white" onClick={handleCreateActivity}>
-//               Create Activity
-//             </Button>
-//           </DialogFooter>
-//         </DialogContent>
-//       </Dialog>
-
-//       {/* Edit Activity Modal */}
-//       <Dialog open={isEditModalOpen} onOpenChange={setIsEditModalOpen}>
-//         <DialogContent className="sm:max-w-[600px] p-0 overflow-hidden rounded-lg">
-//           {currentActivity && (
-//             <>
-//               <DialogHeader className="bg-[#f8f9fa] px-6 py-4 border-b border-gray-200">
-//                 <DialogTitle className="text-xl font-semibold text-[#5C8E77]">Edit Activity</DialogTitle>
-//               </DialogHeader>
-//               <div className="px-6 py-4">
-//                 <Tabs defaultValue="details" className="w-full">
-//                   <TabsList className="grid w-full grid-cols-2 mb-4">
-//                     <TabsTrigger value="details">Activity Details</TabsTrigger>
-//                     <TabsTrigger value="assessment">Assessment</TabsTrigger>
-//                   </TabsList>
-//                   <TabsContent value="details" className="space-y-4">
-//                     <div className="grid gap-4">
-//                       <div className="grid grid-cols-2 gap-4">
-//                         <div className="space-y-2">
-//                           <Label htmlFor="edit-firstName" className="text-[#333]">
-//                             First Name
-//                           </Label>
-//                           <Input
-//                             id="edit-firstName"
-//                             defaultValue={currentActivity.firstName}
-//                             onChange={(e) => setEditFormData({ ...editFormData, firstName: e.target.value })}
-//                             className="border-gray-300"
-//                           />
-//                         </div>
-//                         <div className="space-y-2">
-//                           <Label htmlFor="edit-lastName" className="text-[#333]">
-//                             Last Name
-//                           </Label>
-//                           <Input
-//                             id="edit-lastName"
-//                             defaultValue={currentActivity.lastName}
-//                             onChange={(e) => setEditFormData({ ...editFormData, lastName: e.target.value })}
-//                             className="border-gray-300"
-//                           />
-//                         </div>
-//                       </div>
-//                       <div className="grid grid-cols-2 gap-4">
-//                         <div className="space-y-2">
-//                           <Label htmlFor="edit-chair" className="text-[#333]">
-//                             Chair
-//                           </Label>
-//                           <Select
-//                             defaultValue={currentActivity.chair}
-//                             onValueChange={(value) => setEditFormData({ ...editFormData, chair: value })}
-//                           >
-//                             <SelectTrigger id="edit-chair" className="border-gray-300">
-//                               <SelectValue placeholder={currentActivity.chair} />
-//                             </SelectTrigger>
-//                             <SelectContent>
-//                               <SelectItem value="Chair 03">Chair 03</SelectItem>
-//                               <SelectItem value="Chair 05">Chair 05</SelectItem>
-//                               <SelectItem value="Chair 08">Chair 08</SelectItem>
-//                               <SelectItem value="Chair 10">Chair 10</SelectItem>
-//                               <SelectItem value="Chair 12">Chair 12</SelectItem>
-//                             </SelectContent>
-//                           </Select>
-//                         </div>
-//                         <div className="space-y-2">
-//                           <Label htmlFor="edit-instructor" className="text-[#333]">
-//                             Instructor
-//                           </Label>
-//                           <Select
-//                             defaultValue={currentActivity.instructor}
-//                             onValueChange={(value) => setEditFormData({ ...editFormData, instructor: value })}
-//                           >
-//                             <SelectTrigger id="edit-instructor" className="border-gray-300">
-//                               <SelectValue placeholder={currentActivity.instructor} />
-//                             </SelectTrigger>
-//                             <SelectContent>
-//                               <SelectItem value="Dr. Reyes">Dr. Reyes</SelectItem>
-//                               <SelectItem value="Dr. Mendoza">Dr. Mendoza</SelectItem>
-//                               <SelectItem value="Dr. Santos">Dr. Santos</SelectItem>
-//                               <SelectItem value="Dr. Tan">Dr. Tan</SelectItem>
-//                             </SelectContent>
-//                           </Select>
-//                         </div>
-//                       </div>
-//                       <div className="space-y-2">
-//                         <Label htmlFor="edit-patient" className="text-[#333]">
-//                           Patient
-//                         </Label>
-//                         <Input
-//                           id="edit-patient"
-//                           defaultValue={currentActivity.patient}
-//                           onChange={(e) => setEditFormData({ ...editFormData, patient: e.target.value })}
-//                           className="border-gray-300"
-//                         />
-//                       </div>
-//                       <div className="space-y-2">
-//                         <Label htmlFor="edit-status" className="text-[#333]">
-//                           Status
-//                         </Label>
-//                         <Select
-//                           defaultValue={currentActivity.status}
-//                           onValueChange={(value: Activity["status"]) =>
-//                             setEditFormData({ ...editFormData, status: value })
-//                           }
-//                         >
-//                           <SelectTrigger id="edit-status" className="border-gray-300">
-//                             <SelectValue placeholder={currentActivity.status} />
-//                           </SelectTrigger>
-//                           <SelectContent>
-//                             <SelectItem value="Not started">Not started</SelectItem>
-//                             <SelectItem value="Started">Started</SelectItem>
-//                             <SelectItem value="Completed">Completed</SelectItem>
-//                             <SelectItem value="Incomplete">Incomplete</SelectItem>
-//                           </SelectContent>
-//                         </Select>
-//                       </div>
-//                     </div>
-//                   </TabsContent>
-//                   <TabsContent value="assessment" className="space-y-4">
-//                     <div className="grid gap-4">
-//                       <div className="grid grid-cols-2 gap-4">
-//                         <div className="space-y-2">
-//                           <Label htmlFor="activity-id" className="text-[#333]">
-//                             Activity ID
-//                           </Label>
-//                           <Input
-//                             id="activity-id"
-//                             value={currentActivity.id}
-//                             readOnly
-//                             className="border-gray-300 bg-gray-50"
-//                           />
-//                         </div>
-//                         <div className="space-y-2">
-//                           <Label htmlFor="assessment-procedure" className="text-[#333]">
-//                             Procedure
-//                           </Label>
-//                           <Input
-//                             id="assessment-procedure"
-//                             value={currentActivity.procedure}
-//                             readOnly
-//                             className="border-gray-300 bg-gray-50"
-//                           />
-//                         </div>
-//                       </div>
-//                       <div className="space-y-2">
-//                         <Label htmlFor="assessment-patient" className="text-[#333]">
-//                           Patient
-//                         </Label>
-//                         <Input
-//                           id="assessment-patient"
-//                           value={currentActivity.patient}
-//                           readOnly
-//                           className="border-gray-300 bg-gray-50"
-//                         />
-//                       </div>
-//                       <div className="grid grid-cols-2 gap-4">
-//                         <div className="space-y-2">
-//                           <Label htmlFor="grade" className="text-[#333]">
-//                             Grade
-//                           </Label>
-//                           <Input
-//                             id="grade"
-//                             type="number"
-//                             min="0"
-//                             max="100"
-//                             placeholder="Enter grade (0-100)"
-//                             className="border-gray-300"
-//                             onChange={(e) => setEditFormData({ ...editFormData, grade: e.target.value })}
-//                             defaultValue={currentActivity.grade || ""}
-//                           />
-//                         </div>
-//                         <div className="space-y-2">
-//                           <Label htmlFor="assessment-status" className="text-[#333]">
-//                             Assessment Status
-//                           </Label>
-//                           <Select
-//                             defaultValue={currentActivity.assessmentStatus || "in-progress"}
-//                             onValueChange={(value: "in-progress" | "graded") =>
-//                               setEditFormData({ ...editFormData, assessmentStatus: value })
-//                             }
-//                           >
-//                             <SelectTrigger id="assessment-status" className="border-gray-300">
-//                               <SelectValue placeholder="Select status" />
-//                             </SelectTrigger>
-//                             <SelectContent>
-//                               <SelectItem value="in-progress">In Progress</SelectItem>
-//                               <SelectItem value="graded">Graded</SelectItem>
-//                             </SelectContent>
-//                           </Select>
-//                         </div>
-//                       </div>
-//                       <div className="space-y-2">
-//                         <Label htmlFor="remarks" className="text-[#333]">
-//                           Remarks
-//                         </Label>
-//                         <textarea
-//                           id="remarks"
-//                           rows={4}
-//                           className="w-full p-2 border border-gray-300 rounded-md"
-//                           placeholder="Enter instructor feedback or clinical observations"
-//                           defaultValue={currentActivity.remarks || ""}
-//                           onChange={(e) => setEditFormData({ ...editFormData, remarks: e.target.value })}
-//                         ></textarea>
-//                       </div>
-//                     </div>
-//                   </TabsContent>
-//                 </Tabs>
-//               </div>
-//               <DialogFooter className="bg-[#f8f9fa] px-6 py-4 border-t border-gray-200">
-//                 <Button variant="outline" onClick={() => setIsEditModalOpen(false)} className="border-gray-300">
-//                   Cancel
-//                 </Button>
-//                 <Button className="bg-[#5C8E77] hover:bg-[#406E58] text-white" onClick={() => handleUpdateActivity()}>
-//                   Update Activity
-//                 </Button>
-//               </DialogFooter>
-//             </>
-//           )}
-//         </DialogContent>
-//       </Dialog>
-
-//       {/* Complete Confirmation Modal */}
-//       <Dialog open={isCompleteModalOpen} onOpenChange={setIsCompleteModalOpen}>
-//         <DialogContent className="sm:max-w-[400px] p-0 overflow-hidden rounded-lg">
-//           <DialogHeader className="bg-[#f8f9fa] px-6 py-4 border-b border-gray-200">
-//             <DialogTitle className="text-xl font-semibold text-[#5C8E77]">Confirm Action</DialogTitle>
-//           </DialogHeader>
-//           <div className="px-6 py-4">
-//             <p className="text-[#333]">Are you sure you want to mark this activity as completed?</p>
-//             {activityToAction && (
-//               <div className="mt-3 p-3 bg-[#f8f9fa] rounded-md border border-gray-200">
-//                 <p className="font-medium text-[#333]">{activityToAction.procedure}</p>
-//                 <p className="text-sm text-[#5C8E77]">
-//                   {activityToAction.firstName} {activityToAction.lastName} • {activityToAction.chair}
-//                 </p>
-//                 <p className="text-sm text-gray-500 mt-1">Patient: {activityToAction.patient}</p>
-//               </div>
-//             )}
-//           </div>
-//           <DialogFooter className="bg-[#f8f9fa] px-6 py-4 border-t border-gray-200">
-//             <Button variant="outline" onClick={() => setIsCompleteModalOpen(false)} className="border-gray-300">
-//               Cancel
-//             </Button>
-//             <Button className="bg-blue-600 hover:bg-blue-700 text-white" onClick={handleComplete}>
-//               Confirm
-//             </Button>
-//           </DialogFooter>
-//         </DialogContent>
-//       </Dialog>
-
-//       {/* Archive Confirmation Modal */}
-//       <Dialog open={isArchiveModalOpen} onOpenChange={setIsArchiveModalOpen}>
-//         <DialogContent className="sm:max-w-[400px] p-0 overflow-hidden rounded-lg">
-//           <DialogHeader className="bg-[#f8f9fa] px-6 py-4 border-b border-gray-200">
-//             <DialogTitle className="text-xl font-semibold text-orange-600">Archive Activity</DialogTitle>
-//           </DialogHeader>
-//           <div className="px-6 py-4">
-//             <p className="text-[#333] mb-2">
-//               Are you sure you want to archive this activity? Archived activities will be hidden from the main view but
-//               can be restored later if needed.
-//             </p>
-//             {activityToAction && (
-//               <div className="mt-3 p-3 bg-orange-50 rounded-md border border-orange-200">
-//                 <p className="font-medium text-[#333]">{activityToAction.procedure}</p>
-//                 <p className="text-sm text-orange-700">
-//                   {activityToAction.firstName} {activityToAction.lastName} • {activityToAction.chair}
-//                 </p>
-//                 <p className="text-sm text-gray-500 mt-1">Patient: {activityToAction.patient}</p>
-//               </div>
-//             )}
-//           </div>
-//           <DialogFooter className="bg-[#f8f9fa] px-6 py-4 border-t border-gray-200">
-//             <Button variant="outline" onClick={() => setIsArchiveModalOpen(false)} className="border-gray-300">
-//               Cancel
-//             </Button>
-//             <Button className="bg-orange-600 hover:bg-orange-700 text-white" onClick={handleArchive}>
-//               Archive Activity
-//             </Button>
-//           </DialogFooter>
-//         </DialogContent>
-//       </Dialog>
-//     </div>
-//   )
-// }
